Split `DictArray` and `Column` into separate files
